TypeError: exemption handling for self.capabilities
As mentioned in the homeassistant community thread, some people are getting a "TypeError: argument of type 'NoneType' is not iterable" exemption on line 54.  I debugged the issue and found that it was caused by one of the zwave entities I have is missing a capabilities: attribute.  This patch will allow this error to be handled.